Tweaked damage values for MK221
MK221 DPS equalized with M37 (i.e. decreased), but accuracy increased and spread reduced; idea is to make it a supporting fire weapon that lands slugs with precision for long range support/debuffs/CC 

Also brought sawn-off damage bonus down to relative sanity.

Made it so the UGL on the MK221 is removable.