New admin test-only item: XenoMate HUD
Ok so this is my first PR and I have no idea what I'm doing, so if I accidentally make the admin.txt change go through here please don't gitban me or something.

Now on the actual PR: It adds a new item called the XenoMate HUD that allows humans to view the xenohud that xenomorphs get by default, making it useful for testing in the thunderdome.